+++ source/assocPath-spec.ts
@@ -1,9 +1,32 @@
 import {assocPath} from 'rambda'
 
-describe('R.assocPath', () => {
-  it('happy', () => {
-    const result = assocPath(['b'], 2, {a: 1})
+interface Output{
+  a: number
+  foo: {bar: number}
+}
 
-    result // $ExpectType { a: number; }
+describe('R.assocPath - user must explicitly set type of output', () => {
+  it('with array as path input', () => {
+    const result = assocPath<Output>(['foo','bar'], 2, {a: 1})
+
+    result // $ExpectType Output
+  })
+  it('with string as path input', () => {
+    const result = assocPath<Output>('foo.bar', 2, {a: 1})
+
+    result // $ExpectType Output
+  })
+})
+
+describe('R.assocPath - curried', () => {
+  it('with array as path input', () => {
+    const result = assocPath<Output>(['foo','bar'], 2)({a: 1})
+
+    result // $ExpectType Output
+  })
+  it('with string as path input', () => {
+    const result = assocPath<Output>('foo.bar', 2)({a: 1})
+
+    result // $ExpectType Output
   })
 })